+++ CHANGELOG.md
@@ -4,5 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## v0.1.0
+## [Unreleased]
+
+## [v0.1.3] - 2025-03-31
+- Changed default path to the projects root directory
+- Support for PHP 8.2+ (previously only 8.3+)
+- Improved presets
+
+## [v0.1.2] - 2025-01-27
+- Adds `--text` option to check a string for spelling mistakes
+- Improves base preset
+- Improves laravel preset
+
+## [v0.1.1] - 2025-01-15
+- Ignores "|" character in SpellcheckFormatter
+- Improves base preset
+- Improves laravel preset
+
+## [v0.1.0] - 2025-01-13
 - Adds First Version